move input to atom, decouple from ListItem
